Use different approach for global loading.
Using the same approach as the configmation modal. Having the loading
outside the state allows for greater control and less worrying about
extra renders
